backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
psycopg[binary]>=3.1
python-dotenv>=1.0
pandas>=2.2
openpyxl>=3.1
pytest>=8.0
pytest-django>=4.8
ruff>=0.6
black>=24.0